Split ReactHooks-test into public and internal
